pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "weed"
version = "0.1.0"
description = "AI-powered cannabis strain recommendation system with terpene analysis"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    {name = "Cannabis Recommendation System", email = "dev@example.com"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-ollama>=0.2.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "geopy>=2.4.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "jinja2>=3.1.0",
    "requests>=2.31.0",
    "fastapi-users[sqlalchemy,oauth]==12.1.2",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "httpx-oauth>=0.10.0",
    "python-multipart>=0.0.6",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "setuptools>=65.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/cannabis-strain-recommendations"
Repository = "https://github.com/yourusername/cannabis-strain-recommendations.git"
Issues = "https://github.com/yourusername/cannabis-strain-recommendations/issues"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
